+++ src/ilcdlib/entity/category.py
@@ -13,10 +13,6 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 #
-#  This software was developed with support from the Skanska USA,
-#  Charles Pankow Foundation, Microsoft Sustainability Fund, Interface, MKA Foundation, and others.
-#  Find out more at www.BuildingTransparency.org
-#
 from ilcdlib.common import IlcdXmlReader, NoopBaseReader
 from ilcdlib.dto import Category
 from ilcdlib.xml_parser import T_ET