+++ src/ilcdlib/epd/factory.py
@@ -13,13 +13,10 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 #
-#  This software was developed with support from the Skanska USA,
-#  Charles Pankow Foundation, Microsoft Sustainability Fund, Interface, MKA Foundation, and others.
-#  Find out more at www.BuildingTransparency.org
-#
 from typing import Type
 
 from ilcdlib.epd.dialect.environdec import EnvirondecIlcdXmlEpdReader
+from ilcdlib.epd.dialect.epddenmark import EpdDenmarkIlcdXmlEpdReader
 from ilcdlib.epd.dialect.epditaly import EpdItalyIlcdXmlEpdReader
 from ilcdlib.epd.dialect.epdnorge import EpdNorgeIlcdXmlEpdReader
 from ilcdlib.epd.dialect.indata import IndataIlcdXmlEpdReader
@@ -39,6 +36,7 @@ class EpdReaderFactory:
         "epditaly": EpdItalyIlcdXmlEpdReader,
         "itb": ItbIlcdXmlEpdReader,
         "epdnorge": EpdNorgeIlcdXmlEpdReader,
+        "epddenmark": EpdDenmarkIlcdXmlEpdReader,
     }
     DEFAULT_READER_CLASS = IlcdEpdReader